+++ momaland/learning/continuous_ppo/utils.py
@@ -1,0 +1,27 @@
+"""Wrappers for training.
+
+Parallel only.
+
+TODO AEC.
+"""
+
+import os
+
+import pandas as pd
+
+
+def save_results(returns, exp_name, seed):
+    """Saves the results of an experiment to a csv file.
+
+    Args:
+        returns: a list of triples (timesteps, time, episodic_return)
+        exp_name: experiment name
+        seed: seed of the experiment
+    """
+    if not os.path.exists("results"):
+        os.makedirs("results")
+    filename = f"results/{exp_name}_{seed}.csv"
+    print(f"Saving results to {filename}")
+    df = pd.DataFrame(returns)
+    df.columns = ["Total timesteps", "Time", "Episodic return"]
+    df.to_csv(filename, index=False)
